Removed pro build tag
