+++ .goreleaser.yaml
@@ -71,60 +71,12 @@ builds:
       - -X github.com/loft-sh/vcluster/pkg/telemetry.SyncerVersion={{.Version}}
       - -X github.com/loft-sh/vcluster/pkg/telemetry.telemetryPrivateKey={{.Env.TELEMETRY_PRIVATE_KEY}}
 
-  - id: vcluster-cli-experimental
-    env:
-      - CGO_ENABLED=0
-      - GO111MODULE=on
-    goos:
-      - darwin
-      - linux
-      - windows
-    goarch:
-      - amd64
-      - arm64
-      - arm
-    goarm:
-      - "6"
-    ignore:
-      - goos: darwin
-        goarch: arm
-      - goos: windows
-        goarch: arm
-      - goos: windows
-        goarch: arm64
-    binary: vcluster
-    main: ./cmd/vclusterctl
-    dir: .
-    flags:
-      - -trimpath
-      - -mod
-      - vendor
-    tags:
-      - embed_charts
-      - pro
-    ldflags:
-      - -s -w
-      - -X main.version={{.Version}}
-      - -X github.com/loft-sh/vcluster/pkg/telemetry.SyncerVersion={{.Version}}
-      - -X github.com/loft-sh/vcluster/pkg/telemetry.telemetryPrivateKey={{.Env.TELEMETRY_PRIVATE_KEY}}
-
 archives:
   - id: vcluster_cli_archives
     format: binary
     builds:
       - vcluster-cli
     name_template: "{{ .ProjectName }}-{{ .Os }}-{{ .Arch }}"
-    builds_info:
-      group: root
-      owner: root
-    files:
-      - README.md
-      - LICENSE
-  - id: vcluster_experimental_cli_archives
-    format: binary
-    builds:
-      - vcluster-cli-experimental
-    name_template: "{{ .ProjectName }}-experimental-{{ .Os }}-{{ .Arch }}"
     builds_info:
       group: root
       owner: root
@@ -193,28 +145,6 @@ brews:
       name: loft-bot
       email: 73976634+loft-bot@users.noreply.github.com
     skip_upload: auto
-  - name: vcluster-experimental
-    ids:
-      - vcluster_experimental_cli_archives
-    dependencies:
-      - name: helm
-      - name: kubernetes-cli
-    repository:
-      owner: loft-sh
-      name: homebrew-tap
-    folder: Formula
-    homepage: https://www.vcluster.com
-    license: Apache-2.0
-    description: "Creates fully functional virtual k8s cluster inside host k8s cluster's namespace"
-    conflicts:
-      - vcluster
-      - loft-sh/tap/vcluster
-    test: |
-      help_output = "vcluster root command"
-      assert_match help_output, shell_output("#{bin}/vcluster --help")
-    commit_author:
-      name: loft-bot
-      email: 73976634+loft-bot@users.noreply.github.com
 
 dockers:
   # --- Vcluster images ---
