+++ cmd/vclusterctl/cmd/pro/pro.go
@@ -1,6 +1,3 @@
-//go:build pro
-// +build pro
-
 package pro
 
 import (
