--- pkg/constants/errors.go
@@ -1,7 +1,0 @@
-package constants
-
-import "errors"
-
-var (
-	ErrOnlyInPro = errors.New("this command is only available in vcluster pro")
-)
